+++ examples/config/environment.php
@@ -1,4 +1,16 @@
 <?php
 /**
- * Add any global configs in here
+ * Pre-session modifications go in here
+ *
+ * Obsidian Moon Engine by Dark Prospect Games
+ * An Open Source, Lightweight and 100% Modular Framework in PHP
+ *
+ * PHP version 7
+ *
+ * @category  ObsidianMoonEngine
+ * @package   DarkProspectGames\ObsidianMoonEngine
+ * @author    Alfonso E Martinez, III <opensaurusrex@gmail.com>
+ * @copyright 2011-2018 Dark Prospect Games, LLC
+ * @license   MIT https://darkprospect.net/MIT-License.txt
+ * @link      https://github.com/dark-prospect-games/obsidian-moon-engine/
  */