+++ examples/src/Pages/logerror.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Layout for the app
+ * We caused a log to happen on purpose to show you how it works.
  *
  * Obsidian Moon Engine by Dark Prospect Games
  * An Open Source, Lightweight and 100% Modular Framework in PHP
@@ -14,16 +14,12 @@
  * @license   MIT https://darkprospect.net/MIT-License.txt
  * @link      https://github.com/dark-prospect-games/obsidian-moon-engine/
  */
-?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="author" content="Dark Prospect Games, LLC">
-    <meta name="description" content="An Open Source, Lightweight and 100% Modular Framework in PHP" />
-    <title>Obsidian Moon Engine</title>
-</head>
-<body>
-    <?php echo $content; ?>
-</body>
-</html>
+use DarkProspectGames\ObsidianMoonEngine\Modules\CoreException;
+
+try {
+    $content = 'We have added a log to `logs/errors.log`!';
+    $logger->info('Logger loaded. We successfully saved an info log.');
+    $core->view('layout/layout', compact('content'));
+} catch (CoreException $e) {
+    echo 'An error occurred! Please try refreshing the page.';
+}