+++ examples/src/Views/index.php
@@ -1,6 +1,27 @@
+<?php
+/**
+ * Index page content
+ *
+ * Obsidian Moon Engine by Dark Prospect Games
+ * An Open Source, Lightweight and 100% Modular Framework in PHP
+ *
+ * PHP version 7
+ *
+ * @category  ObsidianMoonEngine
+ * @package   DarkProspectGames\ObsidianMoonEngine
+ * @author    Alfonso E Martinez, III <opensaurusrex@gmail.com>
+ * @copyright 2011-2018 Dark Prospect Games, LLC
+ * @license   MIT https://darkprospect.net/MIT-License.txt
+ * @link      https://github.com/dark-prospect-games/obsidian-moon-engine/
+ */
+?>
 <p>
-    This is an Example App for Obsidian Moon Engine, below you will see content from a module.
+    This is an Example App for Obsidian Moon Engine, below you will see content from
+    a module.
 </p>
 <p>
-    We received a message from DarkProspectGames\ExampleApp\Modules\ExampleModule: "<?=$greeting?>"
+    We received a message from DarkProspectGames\ExampleApp\Modules\ExampleModule:
+    &quot;<?php echo $greeting; ?>&quot;
 </p>
+<p>Click <a href="/missing">here</a> to see a missing page.</p>
+<p>Click <a href="/logerror">here</a> to see saving of logs.</p>